pyproject.toml
[tool.poetry]
name = "debit-scheduler-takehome"
version = "0.1.0"
description = ""
authors = ["Dante Monaldo <dmonaldo@earnup.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^6.2.4"
werkzeug = "^2.0.1"
freezegun = "^1.1.0"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
